+++ custom_components/vaillant_ebus/backend/__init__.py
@@ -6,7 +6,6 @@ from .discovery_service import DiscoveryService
 from .ebus_service import EbusService
 from .entity_factory import EntityDescription, EntityFactoryService
 from .models import DeviceGraph, DeviceNode, DeviceType, EbusdRegister, RegisterMeta, WriteResult
-from .register_service import ParsedValue, RegisterService, RegisterValue, Writeability
 
 __all__ = [
     "DeviceGraph",
@@ -17,10 +16,6 @@ __all__ = [
     "EbusdRegister",
     "EntityDescription",
     "EntityFactoryService",
-    "ParsedValue",
     "RegisterMeta",
-    "RegisterService",
-    "RegisterValue",
     "WriteResult",
-    "Writeability",
 ]